+++ src/main/java/com/example/petshop/config/security/token/TokenConfiguration.java
@@ -1,0 +1,76 @@
+package com.example.petshop.config.security.token;
+
+import com.nimbusds.jose.jwk.JWK;
+import com.nimbusds.jose.jwk.JWKSet;
+import com.nimbusds.jose.jwk.RSAKey;
+import com.nimbusds.jose.jwk.source.ImmutableJWKSet;
+import com.nimbusds.jose.jwk.source.JWKSource;
+import com.nimbusds.jose.proc.SecurityContext;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.oauth2.jwt.JwtDecoder;
+import org.springframework.security.oauth2.jwt.JwtEncoder;
+import org.springframework.security.oauth2.jwt.NimbusJwtDecoder;
+import org.springframework.security.oauth2.jwt.NimbusJwtEncoder;
+import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationProvider;
+import org.springframework.security.provisioning.UserDetailsManager;
+
+@Configuration
+public class TokenConfiguration {
+    final
+    JwtToUserConverter jwtToUserConverter;
+
+    final
+    KeyUtils keyUtils;
+
+    public TokenConfiguration(JwtToUserConverter jwtToUserConverter, KeyUtils keyUtils, UserDetailsManager userDetailsManager) {
+        this.jwtToUserConverter = jwtToUserConverter;
+        this.keyUtils = keyUtils;
+    }
+
+    @Bean
+    @Primary
+    JwtDecoder jwtAccessTokenDecoder() {
+        return NimbusJwtDecoder.withPublicKey(keyUtils.getAccessTokenPublicKey()).build();
+    }
+
+    @Bean
+    @Primary
+    JwtEncoder jwtAccessTokenEncoder() {
+        JWK jwk = new RSAKey
+                .Builder(keyUtils.getAccessTokenPublicKey())
+                .privateKey(keyUtils.getAccessTokenPrivateKey())
+                .build();
+        JWKSource<SecurityContext> jwks = new ImmutableJWKSet<>(new JWKSet(jwk));
+        return new NimbusJwtEncoder(jwks);
+    }
+
+    @Bean
+    @Qualifier("jwtRefreshTokenDecoder")
+    JwtDecoder jwtRefreshTokenDecoder() {
+        return NimbusJwtDecoder.withPublicKey(keyUtils.getRefreshTokenPublicKey()).build();
+    }
+
+    @Bean
+    @Qualifier("jwtRefreshTokenEncoder")
+    JwtEncoder jwtRefreshTokenEncoder() {
+        JWK jwk = new RSAKey
+                .Builder(keyUtils.getRefreshTokenPublicKey())
+                .privateKey(keyUtils.getRefreshTokenPrivateKey())
+                .build();
+        JWKSource<SecurityContext> jwks = new ImmutableJWKSet<>(new JWKSet(jwk));
+        return new NimbusJwtEncoder(jwks);
+    }
+
+    @Bean
+    @Qualifier("jwtRefreshTokenAuthProvider")
+    JwtAuthenticationProvider jwtRefreshTokenAuthProvider() {
+        JwtAuthenticationProvider provider = new JwtAuthenticationProvider(jwtRefreshTokenDecoder());
+        provider.setJwtAuthenticationConverter(jwtToUserConverter);
+        return provider;
+    }
+
+}
